moves ranger.virtualFS -> http/template.mergeFS; simplifies defaults
